split files better

+++ start.js
@@ -13,7 +13,7 @@ import benchmarkOptionsFile from './benchmark-options.js'
 const {resolveAddressesBenchmarkOptions} = benchmarkOptionsFile
 
 const benchmarkNode = (benchmarkFile, benchmarkOptions) => new Promise(resolve => {
-  const benchmarkProcess = spawn('npm', ['run', 'benchmark:node', '--', 'lib/' + benchmarkFile, '--benchmarkOptionsName', benchmarkOptions.name])
+  const benchmarkProcess = spawn('npm', ['run', 'benchmark:node', '--', 'benchmark/' + benchmarkFile, '--benchmarkOptionsName', benchmarkOptions.name])
   benchmarkProcess.stdout.on('data', (data) => process.stdout.write(`${data}`))
   benchmarkProcess.stderr.on('data', (data) => process.stderr.write(`${data}`))
   benchmarkProcess.on('close', (code) => resolve())
@@ -41,17 +41,36 @@ console.log(resolveAddressesBenchmarkOptions)
 const isOnly = (name) => argv.only === name || (argv.only?.length || 0) < 1 || argv.only.includes(name)
 
 // benchmark resolve addresses
-if (isOnly('node')) {
-  for (const benchmarkOptions of resolveAddressesBenchmarkOptions) {
-    fs.removeSync(benchmarkOptions.plebbitOptions.dataPath)
-    const benchmarkFile = 'benchmark-resolve-addresses.js'
-    await benchmarkNode(benchmarkFile, benchmarkOptions)
+if (isOnly('resolve-addresses')) {
+  if (isOnly('node')) {
+    for (const benchmarkOptions of resolveAddressesBenchmarkOptions) {
+      fs.removeSync(benchmarkOptions.plebbitOptions.dataPath)
+      const benchmarkFile = 'benchmark-resolve-addresses.js'
+      await benchmarkNode(benchmarkFile, benchmarkOptions)
+    }
+  }
+  if (isOnly('chrome')) {
+    for (const benchmarkOptions of resolveAddressesBenchmarkOptions) {
+      const benchmarkFile = 'benchmark-resolve-addresses.js'
+      await benchmarkChrome(benchmarkFile, benchmarkOptions)
+    }
   }
 }
-if (isOnly('chrome')) {
-  for (const benchmarkOptions of resolveAddressesBenchmarkOptions) {
-    const benchmarkFile = 'benchmark-resolve-addresses.js'
-    await benchmarkChrome(benchmarkFile, benchmarkOptions)
+
+// benchmark fetch ipns
+if (isOnly('fetch-ipns')) {
+  if (isOnly('node')) {
+    for (const benchmarkOptions of resolveAddressesBenchmarkOptions) {
+      fs.removeSync(benchmarkOptions.plebbitOptions.dataPath)
+      const benchmarkFile = 'benchmark-fetch-ipns.js'
+      await benchmarkNode(benchmarkFile, benchmarkOptions)
+    }
+  }
+  if (isOnly('chrome')) {
+    for (const benchmarkOptions of resolveAddressesBenchmarkOptions) {
+      const benchmarkFile = 'benchmark-fetch-ipns.js'
+      await benchmarkChrome(benchmarkFile, benchmarkOptions)
+    }
   }
 }
